allow disabling backward propagation completely
