+++ application/src/main/java/org/opentripplanner/standalone/config/routerconfig/updaters/MqttGtfsRealtimeUpdaterConfig.java
@@ -27,6 +27,10 @@ public class MqttGtfsRealtimeUpdaterConfig {
         .summary("How backwards propagation should be handled.")
         .description(
           """
+            NONE:
+            Do not propagate delays backwards. Reject real-time updates if the times are not specified
+            from the beginning of the trip.
+
             REQUIRED_NO_DATA:
             Default value. Only propagates delays backwards when it is required to ensure that the times
             are increasing, and it sets the NO_DATA flag on the stops so these automatically updated times
