+++ application/src/main/java/org/opentripplanner/standalone/config/routerconfig/updaters/PollingTripUpdaterConfig.java
@@ -41,6 +41,10 @@ public class PollingTripUpdaterConfig {
         .summary("How backwards propagation should be handled.")
         .description(
           """
+            NONE:
+            Do not propagate delays backwards. Reject real-time updates if the times are not specified
+            from the beginning of the trip.
+
             REQUIRED_NO_DATA:
             Default value. Only propagates delays backwards when it is required to ensure that the times
             are increasing, and it sets the NO_DATA flag on the stops so these automatically updated times
@@ -50,7 +54,7 @@ public class PollingTripUpdaterConfig {
             Only propagates delays backwards when it is required to ensure that the times are increasing.
             The updated times are exposed through APIs.
 
-            ALWAYS
+            ALWAYS:
             Propagates delays backwards on stops with no estimates regardless if it's required or not.
             The updated times are exposed through APIs.
           """
