+++ application/src/main/java/org/opentripplanner/updater/trip/gtfs/BackwardsDelayPropagationType.java
@@ -5,6 +5,12 @@ package org.opentripplanner.updater.trip.gtfs;
  */
 public enum BackwardsDelayPropagationType {
   /**
+   * Do not propagate delays backwards. Reject real-time updates if the times are not specified
+   * from the beginning of the trip.
+   */
+  NONE,
+
+  /**
    * Default value. Only propagates delays backwards when it is required to ensure that the times
    * are increasing, and it sets the NO_DATA flag on the stops so these automatically updated times
    * are not exposed through APIs.
