+++ application/src/main/java/org/opentripplanner/updater/trip/gtfs/BackwardsDelayPropagator.java
@@ -3,48 +3,18 @@ package org.opentripplanner.updater.trip.gtfs;
 import java.util.OptionalInt;
 import org.opentripplanner.transit.model.timetable.RealTimeTripTimesBuilder;
 
-public abstract class BackwardsDelayPropagator {
-
+interface BackwardsDelayPropagator {
   /**
    * Propagate backwards from the first stop with real-time information
    * @return The first stop position with given time if propagation is done.
    */
-  public OptionalInt propagateBackwards(RealTimeTripTimesBuilder builder) {
-    if (builder.getArrivalDelay(0) != null) {
-      // nothing to propagate
-      return OptionalInt.empty();
-    }
-
-    var firstUpdatedIndex = getFirstUpdatedIndex(builder);
-    fillInMissingTimes(builder, firstUpdatedIndex);
-    return OptionalInt.of(firstUpdatedIndex);
-  }
-
-  protected int getFirstUpdatedIndex(RealTimeTripTimesBuilder builder) {
-    var firstUpdatedIndex = 0;
-    while (
-      builder.getArrivalDelay(firstUpdatedIndex) == null &&
-      builder.getDepartureDelay(firstUpdatedIndex) == null
-    ) {
-      ++firstUpdatedIndex;
-      if (firstUpdatedIndex == builder.numberOfStops()) {
-        throw new IllegalArgumentException(
-          "No real-time updates exist in the builder, can't propagate backwards."
-        );
-      }
-    }
-    return firstUpdatedIndex;
-  }
-
-  protected abstract void fillInMissingTimes(
-    RealTimeTripTimesBuilder builder,
-    int firstUpdatedIndex
-  );
+  OptionalInt propagateBackwards(RealTimeTripTimesBuilder builder);
 
   static BackwardsDelayPropagator getBackwardsDelayPropagator(
     BackwardsDelayPropagationType backwardsDelayPropagationType
   ) {
     return switch (backwardsDelayPropagationType) {
+      case NONE -> builder -> OptionalInt.empty();
       case ALWAYS -> new BackwardsDelayAlwaysPropagator();
       case REQUIRED, REQUIRED_NO_DATA -> new BackwardsDelayRequiredPropagator(
         backwardsDelayPropagationType == BackwardsDelayPropagationType.REQUIRED_NO_DATA
